refactor: add subtask register
