+++ backend/helpers/pluginhelper/subtaskmeta_sorter/utils_test.go
@@ -108,3 +108,7 @@ func Test_topologicalSortDifferentElements(t *testing.T) {
 		})
 	}
 }
+
+func Test_SortSubtaskList(t *testing.T) {
+
+}
